pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-usage"
version = "0.1.0"
description = "Claude Code CLI Usage Tracker - Terminal Dashboard"
requires-python = ">=3.10"
dependencies = [
    "textual>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
claude-usage = "claude_usage.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_usage"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
